Se agrega el crossover, se imprime bien la solcuion en el VAR file y se areglan detalles

+++ jmetal-core/src/main/java/org/uma/jmetal/solution/impl/BusProblemStop.java
@@ -31,7 +31,7 @@ public class BusProblemStop {
 		return parada;
 	}
 	
-	public float getOffset(){
+	public int getOffset(){
 		return offset;
 	}
